code/aura-restaurant-system/pi_controller/requirements.txt
# SpeechRecognition
SpeechRecognition
# PyAudio
PyAudio
# pygame
pygame
# google-generativeai
google-generativeai
python-dotenv
gTTS
paho-mqtt
# opencv-python
# deepface
luma.oled
Pillow
smbus2
RPi.GPIO